refactor: Model,Validation을 Domain 하위로 이동
