+++ src/Domain/Model/Race.js
@@ -1,5 +1,5 @@
-import { CAR_MOVE_FORWARD, MAX, MIN } from '../Constants/rules.js';
-import { getRandomNumber } from '../Utils/math.js';
+import { CAR_MOVE_FORWARD, MAX, MIN } from '../../Constants/rules.js';
+import { getRandomNumber } from '../../Utils/math.js';
 
 export default class Race {
   #cars;
