+++ src/app.js
@@ -1,10 +1,10 @@
-import Car from './Model/Car.js';
+import Car from './Domain/Model/Car.js';
 import outputView from './View/output.js';
 import { getCarsName, getGameCount } from './View/input.js';
 import { readUserInputUntilSuccess, FORMATTER } from './View/utils.js';
-import Race from './Model/Race.js';
-import { validateCarsName } from './Validation/carName.js';
-import { validateGameCount } from './Validation/gameCount.js';
+import Race from './Domain/Model/Race.js';
+import { validateCarsName } from './Domain/Validation/carName.js';
+import { validateGameCount } from './Domain/Validation/gameCount.js';
 
 class App {
   async #initialize() {
